requirements.txt
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
honcho==1.0.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-envcfg==0.2.0
SQLAlchemy==1.2.18
Werkzeug==0.14.1
